requirements.txt
pyserial==3.4
pynmeagps==1.0.20
pynmea2==1.19.0
setuptools==57.5.0
PyBluez==0.22
RPi.GPIO==0.7.1
keyboard==0.13.5
